+++ Metroid/Metroid/PlayerOnObstacleHandler.cpp
@@ -1,6 +1,6 @@
 #include "PlayerOnObstacleHandler.h"
 #include "GameObject.h"
 
-void PlayerOnObstacleHandler::Execute(GameObject* obj, float f_x, float f_y) {
+void PlayerOnObstacleHandler::Execute(GameObject* obj, int f_x, int f_y) {
 	obj->physics->ApplyForce(f_x, f_y);
 }
